--- a/src/main/java/com/amazon/jenkins/ec2fleet/CloudNanny.java
+++ b/src/main/java/com/amazon/jenkins/ec2fleet/CloudNanny.java
@@ -3,14 +3,12 @@ package com.amazon.jenkins.ec2fleet;
 import com.google.common.annotations.VisibleForTesting;
 import hudson.Extension;
 import hudson.model.PeriodicWork;
-import hudson.model.Queue;
 import hudson.slaves.Cloud;
 import hudson.widgets.Widget;
 import jenkins.model.Jenkins;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.Callable;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -38,15 +36,10 @@ public class CloudNanny extends PeriodicWork {
 
             try {
                 // Update the cluster states
-                info.add(Queue.withLock(new Callable<EC2FleetStatusInfo>() {
-                    @Override
-                    public EC2FleetStatusInfo call() {
-                        final FleetStateStats stats = fleetCloud.update();
-                        return new EC2FleetStatusInfo(
-                                fleetCloud.getFleet(), stats.getState(), fleetCloud.getLabelString(),
-                                stats.getNumActive(), stats.getNumDesired());
-                    }
-                }));
+                final FleetStateStats stats = fleetCloud.update();
+                info.add(new EC2FleetStatusInfo(
+                        fleetCloud.getFleet(), stats.getState(), fleetCloud.getLabelString(),
+                        stats.getNumActive(), stats.getNumDesired()));
             } catch (Exception e) {
                 // could bad configuration or real exception, we can't do too much here
                 LOGGER.log(Level.INFO, String.format("Error during fleet %s stats update", fleetCloud.name), e);