--- a/src/test/java/com/amazon/jenkins/ec2fleet/EC2FleetCloudTest.java
+++ b/src/test/java/com/amazon/jenkins/ec2fleet/EC2FleetCloudTest.java
@@ -115,6 +115,8 @@ public class EC2FleetCloudTest {
         Registry.setEc2Api(new EC2Api());
     }
 
+    // todo test that we don't remove nodes beyong min size
+
     @Test
     public void provision_fleetIsEmpty() {
         // given