--- a/src/test/java/com/amazon/jenkins/ec2fleet/IdleRetentionStrategyTest.java
+++ b/src/test/java/com/amazon/jenkins/ec2fleet/IdleRetentionStrategyTest.java
@@ -137,4 +137,6 @@ public class IdleRetentionStrategyTest {
         }
     }
 
+    // todo we do nothing if computer doesn't have node
+
 }