## Motivation

Users reported bad Jenkins performance and responsibility due to Plugin. #128 #123 #55 

Detailed analysis shows the ineffective implementation of the plugin work cycle. Which becomes a bottleneck for fleets with a big amount of instances (reported ```50```) or with intensive usage in combination with other slave providers, like EC2 plugin, etc.

## Done

1. Improve plugin thread model to reduce the amount of work in Jenkins cycle, and move it outside
1. Improve locking model for Jenkins resources, lock only when really required
1. Use batch lock instead of per item.
1. Do all heavy/time-consuming work outside of Jenkins cycle

## Implementation

![workflow (2)](https://user-images.githubusercontent.com/473022/64997544-e7ab4a00-d895-11e9-8703-f5087d944dd5.png)

## Result

1. ```35``` times drop for a time which Jenkins spend on plugin execution. ```10 workers x 30 tasks``` consumes ```45 sec``` versus ```~1600 sec``` for an old version
1. ```10%``` improvement for plugin overall execution time
1. **Significant** reduction for scale factor, now a plugin requires ```+2%``` time for additional ```+100%``` load vs ```+400%``` time for an old version
1. Zero blockers for Jenkins cycle. Jenkins overall performance is not affected by plugin any more
1. Make amount of calls to AWS EC2 API more predictable (and reduce it). Plugins can make no more than 30 API calls per min.

## Testing Scenarios

### Small load 1-2 instances

No visible difference between versions

### One-time huge load ```Max 10 workers with 30 tasks```

With downscale at the end after ```1 min```.

| Metric | Old ```sec``` | New ```sec``` |
| --- | ------------- | --- |
| Schedule Up Time  | 141 | 152 |
| Schedule Down Time | 105 | 80 |
| Upscale Method Time | 20 | 0 |
| Downscale Method Time | 827 | 0 |
| Update Method Time | 870 | 45 |

### One-time huge load ```Max 20 workers with 60 tasks```

With downscale at the end after ```1 min```.

| Metric | Old ```sec``` | New ```sec``` |
| --- | ------------- | --- |
| Schedule Up Time  | 150 | 152 |
| Schedule Down Time | 180 | 120 |
| Upscale Method Time | 68 | 0 |
| Downscale Method Time | 3786 | 0 |
| Update Method Time | 3861 | 45 |

### Raw performance data

Performance test ```ProvisionPerformanceTest.java```

```
Old:
30 up in 141 sec, ideal time is 90 sec, overhead is 51 sec
30 down in 105 sec, ideal time is 60 sec, overhead is 45 sec
30 completed in 246 sec, ideal time is 150 sec, overhead is 96 sec
provision meter, total 20, sec count 12 avg 1722 msec
remove meter, total 827, sec count 35 avg 23632 msec
update meter, total 871, sec count 71 avg 12267 msec

New:
30 up in 152 sec, ideal time is 90 sec, overhead is 62 sec
30 down in 80 sec, ideal time is 60 sec, overhead is 20 sec
30 completed in 232 sec, ideal time is 150 sec, overhead is 82 sec
provision meter, total 0, sec count 6 avg 0 msec
remove meter, total 0, sec count 45 avg 0 msec
update meter, total 45, sec count 29 avg 1583 msec

Old:                                                                           
60 up in 150 sec, ideal time is 90 sec, overhead is 60 sec                     
60 down in 180 sec, ideal time is 60 sec, overhead is 120 sec                  
60 completed in 330 sec, ideal time is 150 sec, overhead is 180 sec
provision meter, total 68, sec count 42 avg 1633 msec                          
remove meter, total 3786, sec count 75 avg 50487 msec                          
update meter, total 3861, sec count 139 avg 27780 msec                         

New:                                                                           
60 up in 156 sec, ideal time is 90 sec, overhead is 66 sec                     
60 down in 120 sec, ideal time is 60 sec, overhead is 60 sec                   
60 completed in 276 sec, ideal time is 150 sec, overhead is 126 sec            
provision meter, total 0, sec count 13 avg 0 msec                              
remove meter, total 0, sec count 210 avg 0 msec                                
update meter, total 45, sec count 28 avg 1623 msec                             
```